Moved the enpoint from allele to Variant
