+++ src/main/java/org/alliancegenome/curation_api/controllers/document/AlleleDocumentController.java
@@ -48,9 +48,4 @@ public class AlleleDocumentController implements AlleleDocumentInterface {
 		return ret;
 	}
 
-	@Override
-	public ObjectListResponse<String> getAllVariantNames() {
-		return alleleService.findAllVariantNames();
-	}
-
 }
