+++ src/main/java/org/alliancegenome/curation_api/controllers/document/VariantDocumentController.java
@@ -12,6 +12,7 @@ import org.alliancegenome.curation_api.model.document.es.VariantSummaryDocument;
 import org.alliancegenome.curation_api.model.entities.PredictedVariantConsequence;
 import org.alliancegenome.curation_api.model.entities.Variant;
 import org.alliancegenome.curation_api.model.input.Pagination;
+import org.alliancegenome.curation_api.response.ObjectListResponse;
 import org.alliancegenome.curation_api.response.SearchResponse;
 import org.alliancegenome.curation_api.services.VariantService;
 
@@ -61,5 +62,10 @@ public class VariantDocumentController implements VariantDocumentInterface {
 		SearchResponse<VariantSummaryDocument> ret = new SearchResponse<>(list);
 		ret.setTotalResults(resp.getTotalResults());
 		return ret;
+	}
+
+	@Override
+	public ObjectListResponse<String> getAllVariantNames() {
+		return service.findAllVariantNames();
 	}
 }
