+++ src/main/java/org/alliancegenome/curation_api/interfaces/document/VariantDocumentInterface.java
@@ -3,6 +3,7 @@ package org.alliancegenome.curation_api.interfaces.document;
 import java.util.HashMap;
 
 import org.alliancegenome.curation_api.model.document.es.VariantSummaryDocument;
+import org.alliancegenome.curation_api.response.ObjectListResponse;
 import org.alliancegenome.curation_api.response.SearchResponse;
 import org.alliancegenome.curation_api.view.CurationView;
 import org.eclipse.microprofile.openapi.annotations.parameters.RequestBody;
@@ -12,6 +13,7 @@ import com.fasterxml.jackson.annotation.JsonView;
 
 import jakarta.ws.rs.Consumes;
 import jakarta.ws.rs.DefaultValue;
+import jakarta.ws.rs.GET;
 import jakarta.ws.rs.POST;
 import jakarta.ws.rs.Path;
 import jakarta.ws.rs.Produces;
@@ -28,4 +30,8 @@ public interface VariantDocumentInterface {
 	@Path("/documents")
 	@JsonView({CurationView.VariantSummaryDocument.class, CurationView.SequenceSummaryDocument.class})
 	SearchResponse<VariantSummaryDocument> findDocuments(@DefaultValue("0") @QueryParam("page") Integer page, @DefaultValue("10") @QueryParam("limit") Integer limit, @RequestBody HashMap<String, Object> params);
+
+	@GET
+	@Path("/allVariantNames")
+	ObjectListResponse<String> getAllVariantNames();
 }
