+++ src/main/java/org/alliancegenome/curation_api/services/AlleleService.java
@@ -166,7 +166,4 @@ public class AlleleService extends SubmittedObjectCrudService<Allele, AlleleDTO,
 		return alleleDAO.findAllelesForSummaryByIds(ids);
 	}
 
-	public ObjectListResponse<String> findAllVariantNames() {
-		return alleleDAO.getAllVariantNames();
-	}
 }
