+++ src/main/java/org/alliancegenome/curation_api/services/VariantService.java
@@ -15,6 +15,7 @@ import org.alliancegenome.curation_api.exceptions.ValidationException;
 import org.alliancegenome.curation_api.model.entities.Note;
 import org.alliancegenome.curation_api.model.entities.Variant;
 import org.alliancegenome.curation_api.model.ingest.dto.VariantDTO;
+import org.alliancegenome.curation_api.response.ObjectListResponse;
 import org.alliancegenome.curation_api.response.ObjectResponse;
 import org.alliancegenome.curation_api.services.base.SubmittedObjectCrudService;
 import org.alliancegenome.curation_api.services.validation.VariantValidator;
@@ -114,6 +115,10 @@ public class VariantService extends SubmittedObjectCrudService<Variant, VariantD
 		List<Long> ids = variantDAO.findIdsByParams(params);
 		ids.removeIf(Objects::isNull);
 		return ids;
+	}
+
+	public ObjectListResponse<String> findAllVariantNames() {
+		return variantDAO.getAllVariantNames();
 	}
 
 }
